\name{rstanrt-package}
\alias{rstanrt-package}
\alias{rstanrt}
\docType{package}
\title{
\packageTitle{rstanrt}
}
\description{
\packageDescription{rstanrt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rstanrt}
\packageIndices{rstanrt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rstanrt}

Maintainer: \packageMaintainer{rstanrt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
